# Teqit.Api/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Teqit.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/airplanes": {
      "get": {
        "tags": [
          "Airplanes"
        ],
        "operationId": "airplanes_getAirplanes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Airplane"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Airplane"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Airplane"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Airplanes"
        ],
        "operationId": "airplanes_createAirplane",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/airplanes/{id}": {
      "get": {
        "tags": [
          "Airplanes"
        ],
        "operationId": "airplanes_getAirplaneById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Airplane"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Airplane"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Airplane"
                }
              }
            }
          }
        }
      }
    },
    "/helicopters": {
      "get": {
        "tags": [
          "Helicopters"
        ],
        "operationId": "helicopters_getHelicopters",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Helicopter"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Helicopter"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Helicopter"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Helicopters"
        ],
        "operationId": "helicopters_createHelicopter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/helicopters/{id}": {
      "get": {
        "tags": [
          "Helicopters"
        ],
        "operationId": "helicopters_getHelicopterById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Helicopter"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Helicopter"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Helicopter"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Airplane": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "Helicopter": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Airplanes"
    },
    {
      "name": "Helicopters"
    }
  ]
}